vision_code/requirements.txt
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scipy==1.11.1
scikit-learn==1.3.0
PyYAML==6.0.1
tqdm==4.66.1
Pillow==10.0.0
seaborn==0.12.2